LPS-54271 migrate from monitoring.statistics to monitoring package
