+++ modules/portal/portal-monitoring/src/com/liferay/portal/monitoring/internal/statistics/portlet/PortletStatistics.java
@@ -14,10 +14,10 @@
 
 package com.liferay.portal.monitoring.internal.statistics.portlet;
 
+import com.liferay.portal.kernel.monitoring.DataSampleProcessor;
 import com.liferay.portal.kernel.monitoring.MonitoringException;
+import com.liferay.portal.kernel.monitoring.PortletRequestType;
 import com.liferay.portal.kernel.monitoring.RequestStatus;
-import com.liferay.portal.kernel.monitoring.statistics.DataSampleProcessor;
-import com.liferay.portal.kernel.monitoring.statistics.PortletRequestType;
 import com.liferay.portal.monitoring.internal.statistics.RequestStatistics;
 
 import java.util.HashMap;
