+++ modules/portal/portal-monitoring/src/com/liferay/portal/monitoring/internal/statistics/service/ServiceStatistics.java
@@ -14,9 +14,9 @@
 
 package com.liferay.portal.monitoring.internal.statistics.service;
 
+import com.liferay.portal.kernel.monitoring.DataSampleProcessor;
 import com.liferay.portal.kernel.monitoring.MethodSignature;
 import com.liferay.portal.kernel.monitoring.RequestStatus;
-import com.liferay.portal.kernel.monitoring.statistics.DataSampleProcessor;
 import com.liferay.portal.monitoring.internal.statistics.RequestStatistics;
 
 import java.util.Map;
